core/ui: move window+plugin into core/ and consolidate widgets under ui/
