+++ panel-plugin/core/plugin.h
