+++ panel-plugin/core/register-plugin.c
