+++ panel-plugin/core/register-plugin.h
