+++ panel-plugin/core/resizer.h
