+++ panel-plugin/core/window-geometry.cpp
