+++ panel-plugin/core/window-geometry.h
