+++ panel-plugin/core/window-pages.cpp
