+++ panel-plugin/core/window-pages.h
