+++ panel-plugin/core/window.h
