+++ panel-plugin/meson.build
@@ -5,29 +5,29 @@ plugin_sources = [
   'launcher/category.cpp',
   'launcher/category-button.cpp',
   'launcher/command.cpp',
-  'command-edit.cpp',
+  'ui/command-edit.cpp',
   'launcher/element.cpp',
   'launcher/favorites-page.cpp',
   'places/favourites-section.cpp',
   'places/history-section.cpp',
   'places/home-search-worker.cpp',
   'places/home-section.cpp',
-  'icon-renderer.cpp',
-  'icon-size.cpp',
-  'image-menu-item.h',
+  'ui/icon-renderer.cpp',
+  'ui/icon-size.cpp',
+  'ui/image-menu-item.h',
   'launcher/launcher.cpp',
-  'launcher-icon-view.cpp',
-  'launcher-tree-view.cpp',
+  'ui/launcher-icon-view.cpp',
+  'ui/launcher-tree-view.cpp',
   'migration.cpp',
   'launcher/page.cpp',
   'places/places-item.cpp',
   'places/places-page.cpp',
-  'plugin.cpp',
+  'core/plugin.cpp',
   'profile.cpp',
   'search/query.cpp',
   'launcher/recent-page.cpp',
-  'register-plugin.c',
-  'resizer.cpp',
+  'core/register-plugin.c',
+  'core/resizer.cpp',
   'search/run-action.cpp',
   'search/search-action.cpp',
   'config/usage-stats.cpp',
@@ -45,10 +45,10 @@ plugin_sources = [
   'ui/properties/results-view.cpp',
   'ui/properties/sidebar.cpp',
   'ui/properties/places.cpp',
-  'slot.h',
-  'window.cpp',
-  'window-geometry.cpp',
-  'window-pages.cpp',
+  'ui/slot.h',
+  'core/window.cpp',
+  'core/window-geometry.cpp',
+  'core/window-pages.cpp',
   xfce_revision_h,
   meowmenu_version_h,
 ]
