+++ panel-plugin/settings-dialog.cpp
@@ -19,14 +19,14 @@
 
 #include "launcher/applications-page.h"
 #include "launcher/command.h"
-#include "command-edit.h"
-#include "icon-size.h"
+#include "ui/command-edit.h"
+#include "ui/icon-size.h"
 #include "launcher/launcher.h"
-#include "plugin.h"
+#include "core/plugin.h"
 #include "search/search-action.h"
 #include "settings.h"
-#include "slot.h"
-#include "window.h"
+#include "ui/slot.h"
+#include "core/window.h"
 #include "presets/preset.h"
 #include "presets/preset-io.h"
 #include "search/unified-bar.h"
