+++ panel-plugin/settings.h
@@ -18,7 +18,7 @@
 #ifndef WHISKERMENU_SETTINGS_H
 #define WHISKERMENU_SETTINGS_H
 
-#include "icon-size.h"
+#include "ui/icon-size.h"
 #include "config/usage-stats.h"
 
 #include <string>
