+++ panel-plugin/ui/command-edit.cpp
