+++ panel-plugin/ui/command-edit.h
