+++ panel-plugin/ui/icon-renderer.cpp
