+++ panel-plugin/ui/icon-renderer.h
