+++ panel-plugin/ui/icon-size.cpp
