+++ panel-plugin/ui/icon-size.h
