+++ panel-plugin/ui/image-menu-item.h
