+++ panel-plugin/ui/launcher-icon-view.cpp
