+++ panel-plugin/ui/launcher-icon-view.h
