+++ panel-plugin/ui/launcher-tree-view.cpp
