+++ panel-plugin/ui/launcher-tree-view.h
