+++ panel-plugin/ui/launcher-view.h
