+++ panel-plugin/ui/properties/results-view.cpp
@@ -19,10 +19,10 @@
 
 #include "ui/properties/common.h"
 
-#include "icon-size.h"
-#include "plugin.h"
+#include "ui/icon-size.h"
+#include "core/plugin.h"
 #include "settings.h"
-#include "slot.h"
+#include "ui/slot.h"
 
 #include <libxfce4ui/libxfce4ui.h>
