+++ panel-plugin/ui/properties/user-session.cpp
@@ -19,10 +19,10 @@
 
 #include "ui/properties/common.h"
 
-#include "command-edit.h"
-#include "plugin.h"
+#include "ui/command-edit.h"
+#include "core/plugin.h"
 #include "settings.h"
-#include "slot.h"
+#include "ui/slot.h"
 
 #include <libxfce4ui/libxfce4ui.h>
