+++ panel-plugin/ui/slot.h
