+++ po/POTFILES
@@ -1,30 +1,30 @@
 panel-plugin/launcher/applications-page.cpp
 panel-plugin/launcher/category-button.cpp
 panel-plugin/launcher/category.cpp
-panel-plugin/command-edit.cpp
+panel-plugin/ui/command-edit.cpp
 panel-plugin/launcher/command.cpp
 panel-plugin/launcher/element.cpp
 panel-plugin/launcher/favorites-page.cpp
 panel-plugin/places/favourites-section.cpp
 panel-plugin/places/history-section.cpp
 panel-plugin/places/home-section.cpp
-panel-plugin/icon-renderer.cpp
-panel-plugin/icon-size.cpp
-panel-plugin/launcher-icon-view.cpp
-panel-plugin/launcher-tree-view.cpp
+panel-plugin/ui/icon-renderer.cpp
+panel-plugin/ui/icon-size.cpp
+panel-plugin/ui/launcher-icon-view.cpp
+panel-plugin/ui/launcher-tree-view.cpp
 panel-plugin/launcher/launcher.cpp
 panel-plugin/launcher/page.cpp
 panel-plugin/places/places-item.cpp
 panel-plugin/places/places-page.cpp
-panel-plugin/plugin.cpp
-panel-plugin/preset.cpp
+panel-plugin/core/plugin.cpp
+panel-plugin/presets/preset.cpp
 panel-plugin/profile.cpp
-panel-plugin/query.cpp
+panel-plugin/search/query.cpp
 panel-plugin/launcher/recent-page.cpp
-panel-plugin/resizer.cpp
-panel-plugin/run-action.cpp
-panel-plugin/search-action.cpp
-panel-plugin/search-page.cpp
+panel-plugin/core/resizer.cpp
+panel-plugin/search/run-action.cpp
+panel-plugin/search/search-action.cpp
+panel-plugin/search/search-page.cpp
 panel-plugin/settings-dialog.cpp
 panel-plugin/ui/properties/general.cpp
 panel-plugin/ui/properties/search-bar.cpp
@@ -32,7 +32,7 @@ panel-plugin/ui/properties/search-bar-aliases.cpp
 panel-plugin/ui/properties/search-bar-actions.cpp
 panel-plugin/ui/properties/user-session.cpp
 panel-plugin/settings.cpp
-panel-plugin/window.cpp
+panel-plugin/core/window.cpp
 panel-plugin/core/xfce4-popup-meowmenu.cpp
 
 panel-plugin/meowmenu.desktop.in
